+++ src/SprykerCommunity/Glue/WebhookProcessor/WebhookProcessorConfig.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
 namespace SprykerCommunity\Glue\WebhookProcessor;
 
 use Spryker\Glue\Kernel\AbstractBundleConfig;
+use Spryker\Shared\StorageRedis\StorageRedisConstants;
 use SprykerCommunity\Shared\WebhookProcessor\WebhookProcessorConstants;
 
 class WebhookProcessorConfig extends AbstractBundleConfig
@@ -23,6 +24,11 @@ class WebhookProcessorConfig extends AbstractBundleConfig
      * @var string
      */
     protected const string CONTROLLER_NAME = 'webhook-processor-resource';
+
+    /**
+     * @var string
+     */
+    protected const string DEFAULT_REDIS_INBOX_KEY = 'webhook-processor:inbox';
 
     /**
      * @return string
@@ -46,5 +52,45 @@ class WebhookProcessorConfig extends AbstractBundleConfig
     public function getZedRequestTimeout(): int
     {
         return $this->get(WebhookProcessorConstants::WEBHOOK_ZED_TIMEOUT, 0);
+    }
+
+    /**
+     * @return string
+     */
+    public function getRedisHost(): string
+    {
+        return $this->get(StorageRedisConstants::STORAGE_REDIS_HOST, '');
+    }
+
+    /**
+     * @return int
+     */
+    public function getRedisPort(): int
+    {
+        return (int)$this->get(StorageRedisConstants::STORAGE_REDIS_PORT, 6379);
+    }
+
+    /**
+     * @return string
+     */
+    public function getRedisPassword(): string
+    {
+        return (string)$this->get(StorageRedisConstants::STORAGE_REDIS_PASSWORD, '');
+    }
+
+    /**
+     * @return int
+     */
+    public function getRedisDatabase(): int
+    {
+        return (int)$this->get(StorageRedisConstants::STORAGE_REDIS_DATABASE, 1);
+    }
+
+    /**
+     * @return string
+     */
+    public function getRedisInboxKey(): string
+    {
+        return $this->get(WebhookProcessorConstants::WEBHOOK_REDIS_INBOX_KEY, static::DEFAULT_REDIS_INBOX_KEY);
     }
 }