+++ src/SprykerCommunity/Zed/WebhookProcessor/WebhookProcessorConfig.php
@@ -9,7 +9,9 @@ declare(strict_types=1);
 
 namespace SprykerCommunity\Zed\WebhookProcessor;
 
+use Spryker\Shared\StorageRedis\StorageRedisConstants;
 use Spryker\Zed\Kernel\AbstractBundleConfig;
+use SprykerCommunity\Shared\WebhookProcessor\WebhookProcessorConstants;
 
 class WebhookProcessorConfig extends AbstractBundleConfig
 {
@@ -17,6 +19,11 @@ class WebhookProcessorConfig extends AbstractBundleConfig
      * @var string
      */
     public const string DEFAULT_QUEUE_POOL_NAME = 'synchronizationPool';
+
+    /**
+     * @var string
+     */
+    protected const string DEFAULT_REDIS_INBOX_KEY = 'webhook-processor:inbox';
 
     /**
      * Specification:
@@ -46,5 +53,45 @@ class WebhookProcessorConfig extends AbstractBundleConfig
     public function getDefaultQueuePoolName(): string
     {
         return static::DEFAULT_QUEUE_POOL_NAME;
+    }
+
+    /**
+     * @return string
+     */
+    public function getRedisHost(): string
+    {
+        return $this->get(StorageRedisConstants::STORAGE_REDIS_HOST, '');
+    }
+
+    /**
+     * @return int
+     */
+    public function getRedisPort(): int
+    {
+        return (int)$this->get(StorageRedisConstants::STORAGE_REDIS_PORT, 6379);
+    }
+
+    /**
+     * @return string
+     */
+    public function getRedisPassword(): string
+    {
+        return (string)$this->get(StorageRedisConstants::STORAGE_REDIS_PASSWORD, '');
+    }
+
+    /**
+     * @return int
+     */
+    public function getRedisDatabase(): int
+    {
+        return (int)$this->get(StorageRedisConstants::STORAGE_REDIS_DATABASE, 1);
+    }
+
+    /**
+     * @return string
+     */
+    public function getRedisInboxKey(): string
+    {
+        return $this->get(WebhookProcessorConstants::WEBHOOK_REDIS_INBOX_KEY, static::DEFAULT_REDIS_INBOX_KEY);
     }
 }